Refactor rules package
Introduce new rule types to cover category rules. Furthermore the rule set is generic now.
An additional rulebook is available for the categories.
